+++ technology-transaction-service/src/main/java/com/itts/technologytransactionservice/model/TJsCg.java
@@ -146,7 +146,7 @@ public class TJsCg implements Serializable {
 	/**
  	* 活动id
  	*/
-	private Long jshdId;
+	private Integer jshdId;
 	/**
 	*发布成果审核状态(0待提交;1审核中;2通过;3整改;4拒绝)
  	*/
